requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
boto3==1.16.40
botocore==1.19.40
Django==3.1.4
django-filter==2.4.0
django-money==1.2.1
django-storages==1.11
gunicorn==20.0.4
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2-binary==2.8.6
py-moneyed==0.8.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
